# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Todennus - An Open ID Connect and OAuth2 Provider",
        "title": "Todennus API Endpoints",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/callback": {
            "post": {
                "description": "This endpoint is called by the IdP after it validated the user.\nIt notifies to the server about the authentication result (success or failure) and the inforamtion of user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2"
                ],
                "summary": "Authentication Callback Endpoint",
                "parameters": [
                    {
                        "description": "Authentication result",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OAuth2AuthenticationCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully accept the result",
                        "schema": {
                            "$ref": "#/definitions/dto.OAuth2AuthenticationCallbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerNotFoundErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth2/authorize": {
            "get": {
                "description": "The authorization endpoint is used to interact with the resource owner and obtain an authorization grant.\nThis is the entry point for starting an OAuth2 flow, such as Authorization Code or Implicit.",
                "tags": [
                    "OAuth2"
                ],
                "summary": "OAuth2 Authorization Endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The type of response requested, typically 'code' or 'token'.",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The client ID of the application making the authorization request.",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The URI to which the response will be sent after the authorization.",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The scope of the access request. It defines the level of access the application is requesting.",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "An opaque value used by the client to maintain state between the request and callback.",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "303": {
                        "description": "Redirect to client application with authorization code or error"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth2/consent": {
            "get": {
                "description": "This endpoint serves a consent page when the server needs the user consent for client.",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "OAuth2"
                ],
                "summary": "Consent page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization ID",
                        "name": "authorization_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consent page rendered successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "This endpoint updates the consent result of user then redirect back to the oauth2 authorization endpoint.",
                "tags": [
                    "OAuth2"
                ],
                "summary": "Update consent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization ID",
                        "name": "authorization_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The consent result (accepted or denied)",
                        "name": "consent",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "The accepted scopes of user (usually less than the requested scope).",
                        "name": "scope",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "303": {
                        "description": "Redirect back to oauth2 authorization endpoint"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth2/token": {
            "post": {
                "description": "The token endpoint is used to exchange an authorization code, client credentials, or refresh token for an access token and optionally a refresh token.\nThis is part of the OAuth2 flow to grant access tokens to clients.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth2"
                ],
                "summary": "OAuth2 Token Endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The OAuth2 grant type (authorization_code, client_credentials, refresh_token)",
                        "name": "grant_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The authorization code received from the authorize endpoint (required for authorization_code grant type)",
                        "name": "code",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "The redirect URI used in the authorization request (required for authorization_code grant type)",
                        "name": "redirect_uri",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "The client ID of the application",
                        "name": "client_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The client secret of the application",
                        "name": "client_secret",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The refresh token (required for refresh_token grant type)",
                        "name": "refresh_token",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "The scope of the access request (optional, space-separated)",
                        "name": "scope",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated access token",
                        "schema": {
                            "$ref": "#/definitions/dto.OAuth2TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/session/update": {
            "get": {
                "description": "The user will be redirected to this endpoint by the IdP after it sends the authentication result to the server. \u003cbr\u003e\nThis endpoint updates the user session state to `authenticated`, `unauthenticated`, or `failed authentication`.",
                "tags": [
                    "OAuth2"
                ],
                "summary": "Session Update Endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication id",
                        "name": "authentication_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "303": {
                        "description": "Redirect back to oauth2 authorization endpoint"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/standard.SwaggerBadRequestErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.OAuth2AuthenticationCallbackRequest": {
            "type": "object",
            "properties": {
                "authorization_id": {
                    "type": "string",
                    "example": "djG4l..."
                },
                "error": {
                    "type": "string",
                    "example": ""
                },
                "idp_secret": {
                    "type": "string",
                    "example": "Sde3kl..."
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "user_id": {
                    "type": "string",
                    "example": "329780019283901"
                },
                "username": {
                    "type": "string",
                    "example": "huykingsofm"
                }
            }
        },
        "dto.OAuth2AuthenticationCallbackResponse": {
            "type": "object",
            "properties": {
                "authentication_id": {
                    "type": "string",
                    "example": "hlqWe..."
                }
            }
        },
...
        "dto.OAuth2TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
...
        "standard.SwaggerBadRequestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "error_description": {
                    "type": "string",
                    "example": "invalid field: require string but got int"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "standard.SwaggerMetadata": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string",
                    "example": "IIJORWQpIvDMzzNf"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-10-23T13:52:29.459752901+07:00"
                }
            }
        },
        "standard.SwaggerNotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                },
                "error_description": {
                    "type": "string",
                    "example": "not found user with id 323979471029873"
                },
                "metadata": {
                    "$ref": "#/definitions/standard.SwaggerMetadata"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
    }
}